Move people models into their own app
