+++ tbx/people/templates/people/person_index_page.html
